# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitnes Planer API",
        "description": "Dokumentacija koristeći PHP Atribute",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Lokalni API server"
        }
    ],
    "paths": {
...
        "/users/{user}/ciljevi": {
            "get": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Prikazuje ciljeve određenog korisnika",
                "operationId": "afd6d804eed25b20545280c6a4d9c018",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista ciljeva"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Trener dodaje cilj za vežbača",
                "operationId": "48b40a7a329b9636e14e9228ed65965b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cilj kreiran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/cilj": {
            "get": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Prikaz poslednjeg cilja (kalorije) ili liste ciljeva",
                "operationId": "68a72f467446511e3051fab06ecd0869",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Kreiranje novog cilja",
                "operationId": "f50847a800edbe1cabf932b59bd06a5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hidriranost",
                                    "tezina",
                                    "kalorije"
                                ],
                                "properties": {
                                    "hidriranost": {
                                        "type": "number",
                                        "example": 2.5
                                    },
                                    "tezina": {
                                        "type": "number",
                                        "example": 75.5
                                    },
                                    "kalorije": {
                                        "type": "integer",
                                        "example": 2500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cilj uspešno kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/all-ciljevi": {
            "get": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Dobijanje svih ciljeva korisnika",
                "operationId": "8de3ce333dd475bf2f10e96a4abc8f9f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cilj/{id}": {
            "delete": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Brisanje cilja",
                "operationId": "fc7c069414f86b8ebb2e9bf6377fbc89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/hidriranost": {
            "get": {
                "tags": [
                    "Hidriranost"
                ],
                "summary": "Prikaz svih zapisa o hidriranosti",
                "operationId": "2a28dab337ba57f340d06d5db0ac9299",
                "responses": {
                    "200": {
                        "description": "Lista svih zapisa"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hidriranost"
                ],
                "summary": "Kreiranje novog zapisa o hidriranosti za danas",
                "operationId": "c0a61a16c17b815ba82e794c876cde8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ukupno": {
                                        "type": "number",
                                        "example": 0.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zapis kreiran"
                    },
                    "400": {
                        "description": "Zapis za danas već postoji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hidriranost-danas": {
            "get": {
                "tags": [
                    "Hidriranost"
                ],
                "summary": "Provera da li postoji zapis o hidriranosti za današnji dan",
                "operationId": "119c6066fb9b8d5111a3317051cd6519",
                "responses": {
                    "200": {
                        "description": "Informacija o postojanju zapisa"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hidriranost/{hidriranost}": {
            "put": {
                "tags": [
                    "Hidriranost"
                ],
                "summary": "Dodavanje količine vode na današnji zapis",
                "operationId": "8ba006af85d8b55a66e0273f4468f597",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ukupno": {
                                        "type": "number",
                                        "example": 0.25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vrednost ažurirana"
                    },
                    "404": {
                        "description": "Zapis nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}